chatbot/requirements.txt
discord.py
groq
python-dotenv
aiosqlite
